requirements.txt
asgiref==3.2.7
asn1crypto==1.3.0
attrs==19.3.0
backcall==0.1.0
bleach==3.1.0
certifi==2020.6.20
cffi==1.14.0
chardet==3.0.4
colorama==0.4.3
conda==4.8.3
conda-package-handling==1.6.0
cryptography==2.8
decorator==4.4.2
defusedxml==0.6.0
dj-database-url==0.5.0
dj-static==0.0.6
Django==3.0.3
entrypoints==0.3
future==0.18.2
gunicorn==20.0.4
idna==2.9
importlib-metadata==1.5.0
ipykernel==5.1.4
ipython==7.13.0
ipython-genutils==0.2.0
ipywidgets==7.5.1
jedi==0.16.0
Jinja2==2.11.1
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==6.1.0
jupyter-console==6.1.0
jupyter-core==4.6.1
line-bot-sdk==1.16.0
MarkupSafe==1.1.1
menuinst==1.4.16
mistune==0.8.4
nbconvert==5.6.1
nbformat==5.0.4
notebook==6.0.3
pandocfilters==1.4.2
parso==0.6.2
pickleshare==0.7.5
prometheus-client==0.7.1
prompt-toolkit==3.0.3
psycopg2==2.8.5
pycosat==0.6.3
pycparser==2.20
Pygments==2.6.1
pyOpenSSL==19.1.0
pyrsistent==0.15.7
PySocks==1.7.1
python-dateutil==2.8.1
pytz==2020.1
pywin32==227
pywinpty==0.5.7
pyzmq==18.1.1
qtconsole==4.7.1
QtPy==1.9.0
requests==2.23.0
ruamel-yaml==0.15.87
Send2Trash==1.5.0
six==1.14.0
sqlparse==0.3.1
static3==0.7.0
terminado==0.8.3
testpath==0.4.4
tornado==6.0.4
tqdm==4.43.0
traitlets==4.3.3
urllib3==1.25.8
wcwidth==0.1.8
webencodings==0.5.1
widgetsnbextension==3.5.1
win-inet-pton==1.1.0
wincertstore==0.2
zipp==2.2.0
